Add comments and update the readme.
I probably should have done this when originally writing this, but I
assumed the project would be viewed alongside a presentation I was
giving.

But I realized that people might be looking at this example weeks from
now and might not remember the things I was talking about.

I also figured there's no reason not to leave this public and wanted to
make it easier for anyone who hasn't seen the presentation to follow my
logic.