Update filtering to be done alongside pagination.
I'm actually not sure I like how I did this, but I'm running out of time
today. It seems like the filter should be a single action and the
pagination a separate one. Also the page change when you change a filter
seems a little out of place, it should probably be in an action as well.